backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pyswisseph==2.10.3.2
pytz==2023.3
timezonefinder==6.2.0
